+++ grease/src/error.rs
@@ -10,6 +10,8 @@ use abi_stable::{
     std_types::{RArc, RBoxError, RVec},
     StableAbi,
 };
+use std::collections::HashSet;
+use std::hash::{Hash, Hasher};
 
 /// An external error.
 pub type ExternalError = dyn std::error::Error + Send + Sync;
@@ -150,21 +152,51 @@ impl Error {
             inner: InnerError::Aborted,
         }
     }
+
+    /// Returns the root source of this error.
+    ///
+    /// Unlike `source()`, this may return this error (if no source is available).
+    ///
+    /// This is the root error traversing known error types, namely nested `Error`s.
+    pub fn root_error(&self) -> Self {
+        let mut inner = &self.inner;
+        let mut src: &dyn std::error::Error = self.error_ref();
+        // Traverse all sources, tracking the last source that was an InnerError.
+        while let Some(nsrc) = src.source() {
+            src = nsrc;
+            if let Some(e) = downcast_ref::<WrappedError>(src) {
+                inner = &e.inner;
+            } else if let Some(e) = downcast_ref::<InnerError>(src) {
+                inner = e;
+            }
+        }
+        Error {
+            inner: inner.clone(),
+        }
+    }
+
+    /// Whether this error is an aggregate error.
+    pub fn is_aggregate(&self) -> bool {
+        if let InnerError::Nested(_) = &self.inner {
+            true
+        } else {
+            false
+        }
+    }
+
+    /// Whether this error is an aborted error.
+    pub fn is_aborted(&self) -> bool {
+        if let InnerError::Aborted = &self.inner {
+            true
+        } else {
+            false
+        }
+    }
 }
 
 impl AsRef<ExternalError> for Error {
     fn as_ref(&self) -> &ExternalError {
         &self.inner
-    }
-}
-
-fn has_inner_error(v: &(dyn std::error::Error + 'static)) -> bool {
-    if v.is::<InnerError>() {
-        true
-    } else if let Some(s) = v.source() {
-        has_inner_error(s)
-    } else {
-        false
     }
 }
 
@@ -176,7 +208,6 @@ where
         let ext: Box<ExternalError> = v.into();
         match ext.downcast::<WrappedError>() {
             Ok(v) => Error { inner: v.inner },
-            // TODO: traverse source() to find WrappedError?
             Err(e) => Error {
                 inner: InnerError::New(RArc::new(RBoxError::from_box(e))),
             },
@@ -219,5 +250,62 @@ impl<T: std::fmt::Display> std::fmt::Display for ErrorContext<T> {
 impl<T: std::fmt::Display + std::fmt::Debug> std::error::Error for ErrorContext<T> {
     fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
         Some(self.err.error_ref())
+    }
+}
+
+#[derive(Debug)]
+struct HashBySource(Error);
+
+impl HashBySource {
+    fn as_ptr(&self) -> *const RBoxError {
+        match &self.0.inner {
+            InnerError::New(a) => a.as_ref() as *const RBoxError,
+            _ => panic!("invalid entry in UniqueErrorSources"),
+        }
+    }
+}
+
+impl Hash for HashBySource {
+    fn hash<H: Hasher>(&self, h: &mut H) {
+        self.as_ptr().hash(h)
+    }
+}
+
+impl PartialEq for HashBySource {
+    fn eq(&self, other: &Self) -> bool {
+        self.as_ptr() == other.as_ptr()
+    }
+}
+
+impl Eq for HashBySource {}
+
+/// A set to track errors with referentially-unique sources.
+#[derive(Default, Debug)]
+pub struct UniqueErrorSources(HashSet<HashBySource>);
+
+impl UniqueErrorSources {
+    /// Create an empty set.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Insert an error into the set.
+    pub fn insert(&mut self, err: Error) -> bool {
+        let err = err.root_error();
+        match &err.inner {
+            InnerError::New(_) => (),
+            _ => return false,
+        }
+        self.0.insert(HashBySource(err))
+    }
+
+    /// Return the number of unique errors in the set.
+    pub fn len(&self) -> usize {
+        self.0.len()
+    }
+
+    /// Iterate over the errors in the set.
+    pub fn iter(&self) -> impl Iterator<Item = &Error> {
+        self.0.iter().map(|h| &h.0)
     }
 }